src/requirements.txt
adal==1.2.7
applicationinsights==0.11.10
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
azure-common==1.1.27
azure-core==1.14.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azure-nspkg==3.0.2
azure-servicebus==7.2.0
azure-storage==0.36.0
azure-storage-blob==12.8.1
azureml==0.2.7
azureml-automl-core==1.29.0
azureml-core==1.29.0
azureml-dataprep==2.15.1
azureml-dataprep-native==33.0.0
azureml-dataprep-rslex==1.13.0
azureml-dataset-runtime
azureml-pipeline==1.29.0
azureml-pipeline-core==1.29.0
azureml-pipeline-steps==1.29.0
azureml-sdk==1.29.0
azureml-telemetry==1.29.0
azureml-train==1.29.0
azureml-train-automl-client==1.29.0
azureml-train-core==1.29.0
azureml-train-restclients-hyperdrive==1.29.0
azureml-widgets==1.29.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.3
distro==1.5.0
docker==4.4.4
docopt==0.6.2
dotnetcore2==2.1.20
entrypoints==0.3
filelock==3.0.12
fusepy==3.0.1
future==0.18.2
h2o==3.26.0.3
huggingface-hub==0.0.8
idna==2.10
importlib-metadata==4.0.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
msal==1.12.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.4.3
notebook==6.4.0
numpy
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.10
portalocker==1.7.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
pyarrow==3.0.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
SecretStorage==3.3.1
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
tabulate==0.8.9
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
transformers==4.6.1
typing-extensions==3.10.0.0
uamqp==1.4.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.0
widgetsnbextension==3.5.1
wincertstore==0.2
yarg==0.1.9
zipp==3.4.1
